add dynamic editText views
