+++ app/src/main/res/layout/main_activity.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black_gray"
-    tools:context=".MainActivity" >
+    tools:context=".ui.main.MainActivity" >
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/createRankBtn"
