+++ app/src/main/res/menu/toolbar_menu.xml
@@ -5,11 +5,15 @@
     <item android:id="@+id/search_icon"
         android:icon="@drawable/ic_search_icon"
         android:title="Search"
-        app:showAsAction="ifRoom|collapseActionView"
+        app:showAsAction="always|collapseActionView"
         app:actionViewClass="androidx.appcompat.widget.SearchView" />
 
     <item android:id="@+id/search_menu_item01"
-        android:title="Option"
+        android:title="Settings"
+        app:showAsAction="never" />
+
+    <item android:id="@+id/search_menu_item02"
+        android:title="Filter"
         app:showAsAction="never" />
 
 </menu>
