+++ app/src/main/res/values/dimens.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<resources>
+<resources xmlns:tools="http://schemas.android.com/tools">
     <!--
     Because the window insets on round devices are larger than 15dp, this padding only applies
     to square screens.
     -->
     <dimen name="box_inset_layout_padding">0dp</dimen>
+    <dimen name="floatingActionButton01" tools:override="true">36dp</dimen>
 
     <!--
     This padding applies to both square and round screens. The total padding between the buttons
